+++ src/containers/MainPage/MainPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ShowGrid from "../../components/ShowGrid/ShowGrid";
+import ShowGrid from "../ShowGrid/ShowGrid";
 import ScrollContainer from "../../components/ScrollContainer/ScrollContainer";
 
 export const MainPage = () => {
