+++ src/containers/SearchBox/searchBox.styles.ts
